+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [0.10.2] - 2021-03
 ### Added
+- New flag '--deduplicate' for `treesapp create` to remove redundant sequences within 99.9% similarity before querying Entrez.
 
 ### Fixed